refactor(searchbar): defineOptions instead of two script
